% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_proper.R
\name{corr_proper}
\alias{corr_proper}
\title{corr_proper}
\usage{
corr_proper(data, columns = 1:length(data), p.adjust = NULL)
}
\arguments{
\item{data}{a dataframe or tibble.}

\item{columns}{can be either index or colnames of data, which enables subsetting data to test.}

\item{p.adjust}{p.adjust receive a string of holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none", which would conduct p value adjust.}
}
\value{
a list. contains the matrix and long form result of correlation matrix
}
\description{
chooses proper correlation method (pearson or spearman) based on normality of the tested variables.
}
\examples{
data("mtcars")

#Use all columns to compute correlaton
corr_proper(mtcars)

#Use only the first 3 columns,and use fdr correction
corr_proper(mtcars, 1:3, 'fdr')

}
