% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrProper.R
\name{corrProper}
\alias{corrProper}
\title{Correlation Analysis with Proper Method}
\usage{
corrProper(data, y = NULL, p.adjust = NULL)
}
\arguments{
\item{data}{a dataframe or tibble.}

\item{y}{the second matrix or dataframe with the same number of rows as data.If y is provided, the correlation just between variables in data and y will be calculated. If y is not provided, pair-wise correlation between variables in data will be calculated.}

\item{p.adjust}{p.adjust receive a string of holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none", which would conduct p value adjust.}
}
\value{
a list. contains the matrix and long form result of correlation matrix:\itemize{
\item\code{r:} the symmetric matrix of correlations
\item\code{p:} two tailed probability of t for each correlation
\item\code{n:} the number of sample used in this pair of correlation
\item\code{method:} the correlation method used in this pair of correlation analysis. If both variables are ordinal, kendall correlation analysis will be used. If both variables are numeric and meet normality, pearson correlation analysis will be used. If either of the two variables doesn't meet normality, spearman correlation analysis will be used. If one variable is ordinal and the other is numerical ,spearman correlation analysis will be used. If either variable is factor, NA will be returned to all results.
\item\code{t:} value of t-test for each correlation
\item\code{se:} standard error of the correlation
\item\code{cor_long:} a tibble, the long format of correlation analysis results, NA results will not be shown}
}
\description{
chooses proper correlation method (pearson or spearman) based on normality of the tested variables.
}
\examples{
data("mtcars")
corrProper(mtcars)

#Use fdr correction
corrProper(mtcars, p.adjust = 'fdr')

#Calculate correlation between two data frames
corrProper(mtcars[1:3],mtcars[4:5], p.adjust = 'fdr')

}
