% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desCompareIndTwoGroups.R
\name{desCompareIndTwoGroups}
\alias{desCompareIndTwoGroups}
\title{desCompareIndTwoGroups}
\usage{
desCompareIndTwoGroups(data, columns, group, ci = FALSE)
}
\arguments{
\item{data}{a data.frame or tibble}

\item{columns}{dependent variables needed to be test, could be either string or column index}

\item{group}{the grouping variable, could be either string or column index}

\item{ci}{a logical. Determine whether the ci of effect size should be calculated. Default is False}
}
\value{
a data frame that contains description and comparison results.
}
\description{
gives a easy-to-read result of two independent group description and comparison analysis.
}
\details{
The comparison is based on the normality and variance quality. If the data don't meet normality, Mann-Whitney test will be conducted. Ohterwise the student t-test would be conducted.
}
\examples{
data("mtcars")
desCompareIndTwoGroups(data = mtcars, columns = -8, group = 8)


}
