% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResiduals.R
\name{getResiduals}
\alias{getResiduals}
\title{Get Residuals with Linear Model}
\usage{
getResiduals(data, cov, addmean = TRUE)
}
\arguments{
\item{data}{a data frame or tibble, contains the columns to calculate residuals}

\item{cov}{a data frame or tibble, contains the columns as covariate. The length of row should be the same as data}

\item{addmean}{logical, should the mean of original data should be added back to the residual, default is true}
}
\value{
a data frame with same dimension with data, containing the residuals of the original columns
}
\description{
get residuals of columns in data regressed cov with \code{\link[stats]{lm}}
}
\examples{
data("mtcars")
getResiduals(mtcars[1:3], mtcars[4:5])

#get residual without adding mean back
getResiduals(mtcars[1:3], mtcars[4:5],addmean = FALSE)

}
