% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sameChoiceDetect.R
\name{sameChoiceDetect}
\alias{sameChoiceDetect}
\title{Detect continuous same choices in the inventory}
\usage{
sameChoiceDetect(data, k = NULL, index = NULL, keep = TRUE)
}
\arguments{
\item{data}{A data frame. Each row represents an observation, and each column represents an item}

\item{k}{The minimum number of continuous same choices to detect.}

\item{index}{A vector of index of column. Used when there is only a subset of columns is to detected.}

\item{keep}{A logical indicating whether keep the information about the continuous same choices. If is set FALSE, the observations with same choices will be discarded and the information will not be kept. Default is TRUE.}
}
\value{
A data frame, a combination of the original data frame and two other columns:
\itemize{
\item{\code{identical}: An integer, indicates the number of  continuous same choices in this observation's answers.}
\item{\code{start_index}: The index of the first index of the detected continuous same choices.}
}
}
\description{
Detect continuous same choices in the inventory
}
\examples{
data('mtcars')
sameChoiceDetect(mtcars, k = 4)

}
